requirements.txt
paper-qa
sentence-transformers
